+++ src/app/routes/connector-ui/asset-page/asset-edit-dialog/form/model/asset-advanced-form-model.ts
@@ -1,7 +1,13 @@
-import {FormControl, ɵFormGroupValue} from '@angular/forms';
+import {
+  FormArray,
+  FormControl,
+  FormGroup,
+  ɵFormGroupValue,
+} from '@angular/forms';
 import {DataCategorySelectItem} from '../../../data-category-select/data-category-select-item';
 import {DataSubcategorySelectItem} from '../../../data-subcategory-select/data-subcategory-select-item';
 import {TransportModeSelectItem} from '../../../transport-mode-select/transport-mode-select-item';
+import {TemporalCoverageFormModel} from './temporal-coverage-form-model';
 
 /**
  * Form Model for AssetEditorDialog > Advanced
@@ -13,6 +19,15 @@ export interface AssetAdvancedFormModel {
   dataModel: FormControl<string>;
   geoReferenceMethod: FormControl<string>;
   transportMode: FormControl<TransportModeSelectItem | null>;
+  sovereignLegalName: FormControl<string>;
+  geoLocation: FormControl<string>;
+  nutsLocations: FormArray<FormControl<string>>;
+  dataSampleUrls: FormArray<FormControl<string>>;
+  referenceFileUrls: FormArray<FormControl<string>>;
+  referenceFilesDescription: FormControl<string>;
+  conditionsForUse: FormControl<string>;
+  dataUpdateFrequency: FormControl<string>;
+  temporalCoverage: FormGroup<TemporalCoverageFormModel>;
 }
 
 /**