+++ src/app/routes/connector-ui/asset-page/asset-page.module.ts
@@ -6,6 +6,7 @@ import {MatButtonModule} from '@angular/material/button';
 import {MatCardModule} from '@angular/material/card';
 import {MatCheckboxModule} from '@angular/material/checkbox';
 import {MatChipsModule} from '@angular/material/chips';
+import {MatNativeDateModule} from '@angular/material/core';
 import {MatDatepickerModule} from '@angular/material/datepicker';
 import {MatDialogModule} from '@angular/material/dialog';
 import {MatDividerModule} from '@angular/material/divider';
@@ -61,6 +62,7 @@ import {TransportModeSelectComponent} from './transport-mode-select/transport-mo
     MatInputModule,
     MatIconModule,
     MatDialogModule,
+    MatNativeDateModule,
 
     // EDC UI Modules
     CatalogModule,